requirements.txt
asgiref==3.6.0
atomicwrites==1.4.1
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
chardet==3.0.4
click==8.1.3
colorama==0.4.6
cssbeautifier==1.14.7
Django==4.1.7
django-ckeditor==6.5.1
django-js-asset==2.0.0
django_debug_toolbar==3.8.1
djlint==1.19.15
EditorConfig==0.12.3
Faker==12.0.1
flake8==6.0.0
html-tag-names==0.1.2
html-void-elements==0.1.0
idna==2.8
jsbeautifier==1.14.7
mccabe==0.7.0
mixer==7.1.2
more-itertools==9.0.0
packaging==23.0
pathspec==0.11.0
Pillow==9.4.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.64.1
tzdata==2022.7
urllib3==1.25.11
wcwidth==0.2.6
